requirements.txt
streamlit
openai
python-dotenv
azure-cosmos